refactor: Move tests to core/tests
