+++ datafusion/core/src/physical_optimizer/mod.rs
@@ -21,9 +21,6 @@
 //! "Repartition" or "Sortedness"
 //!
 //! [`ExecutionPlan`]: crate::physical_plan::ExecutionPlan
-mod tests {
-    pub mod aggregate_statistics;
-}
 pub mod coalesce_batches;
 pub mod combine_partial_final_agg;
 pub mod enforce_distribution;
